@@ -3,8 +3,8 @@
 buildscript {
   ext.versions = [
       minSdk    : 21,
-      compileSdk: 27,
-      supportLib: '27.1.1',
+      compileSdk: 28,
+      androidx  : '1.0.0-rc01',
       timber    : '4.7.0',
       kotlin    : '1.2.60',
   ]
