@@ -1,6 +1,6 @@
 package me.saket.expand.sample
 
-import android.support.v7.util.DiffUtil
+import androidx.recyclerview.widget.DiffUtil
 
 data class EmailThread(
     val sender: Person,
